Make use of underlying kernel matrix in custom kernel
  - The kernel matrix stored inside a custom kernel is always
    float32 type. So, instead of creating a double matrix and
    doing computation on that, let's use the float32 matrix
    itself. Casting the matrix to double doesn't increase the
    accuracy anyway. On the other hand, float32 matrix is of
    half the size, and hence we get significant performance
    boost.

    check
    https://gist.github.com/lambday/66297100d490dab98c2c398ed80978a2

+++ src/shogun/statistical_testing/BTestMMD.cpp
@@ -44,7 +44,7 @@ void CBTestMMD::set_num_blocks_per_burst(index_t num_blocks_per_burst)
 	get_data_manager().set_num_blocks_per_burst(num_blocks_per_burst);
 }
 
-const std::function<float64_t(SGMatrix<float64_t>)> CBTestMMD::get_direct_estimation_method() const
+const std::function<float32_t(SGMatrix<float32_t>)> CBTestMMD::get_direct_estimation_method() const
 {
 	return mmd::WithinBlockDirect();
 }
