+++ src/shogun/statistical_testing/BTestMMD.h
@@ -26,7 +26,7 @@ namespace shogun
 
 class CBTestMMD : public CMMD
 {
-	using operation = std::function<float64_t(SGMatrix<float64_t>)>;
+	using operation=std::function<float32_t(SGMatrix<float32_t>)>;
 public:
 	CBTestMMD();
 	virtual ~CBTestMMD();
