+++ src/shogun/statistical_testing/LinearTimeMMD.h
@@ -26,7 +26,7 @@ namespace shogun
 
 class CLinearTimeMMD : public CMMD
 {
-	using operation = std::function<float64_t(SGMatrix<float64_t>)>;
+	using operation=std::function<float32_t(SGMatrix<float32_t>)>;
 public:
 	CLinearTimeMMD();
 	virtual ~CLinearTimeMMD();
