+++ src/shogun/statistical_testing/internals/mmd/UnbiasedFull.cpp
@@ -29,24 +29,24 @@ UnbiasedFull::UnbiasedFull(index_t n) : n_x(n)
 {
 }
 
-float64_t UnbiasedFull::operator()(SGMatrix<float64_t> km)
+float32_t UnbiasedFull::operator()(SGMatrix<float32_t> km)
 {
-	using MatrixXt = const Eigen::MatrixXd;
-	using Block = const Eigen::Block<Eigen::Map<MatrixXt>>;
+	using MatrixXt=const Eigen::MatrixXf;
+	using Block=const Eigen::Block<Eigen::Map<MatrixXt>>;
 
 	Eigen::Map<MatrixXt> map(km.matrix, km.num_rows, km.num_cols);
-	index_t n_y = km.num_rows - n_x;
+	index_t n_y=km.num_rows-n_x;
 
-	Block& b_x = map.block(0, 0, n_x, n_x);
-	auto term_1 = b_x.sum() - b_x.diagonal().sum();
+	Block& b_x=map.block(0, 0, n_x, n_x);
+	auto term_1=b_x.sum()-b_x.diagonal().sum();
 
-	Block& b_y = map.block(n_x, n_x, n_y, n_y);
-	auto term_2 = b_y.sum() - b_y.diagonal().sum();
+	Block& b_y=map.block(n_x, n_x, n_y, n_y);
+	auto term_2=b_y.sum()-b_y.diagonal().sum();
 
-	Block& b_xy = map.block(n_x, 0, n_y, n_x);
-	auto term_3 = b_xy.sum();
+	Block& b_xy=map.block(n_x, 0, n_y, n_x);
+	auto term_3=b_xy.sum();
 
-	auto statistic = term_1/n_x/(n_x-1) + term_2/n_y/(n_y-1) - 2*term_3/n_x/n_y;
+	auto statistic=term_1/n_x/(n_x-1)+term_2/n_y/(n_y-1)-2*term_3/n_x/n_y;
 
 	return statistic;
